Revert "Track composite"
