+++ pkg/config/config.go
@@ -1,6 +1,7 @@
 package config
 
 import (
+	"fmt"
 	"os"
 
 	"github.com/go-logr/zapr"
@@ -8,7 +9,6 @@ import (
 	"go.uber.org/zap/zapcore"
 	"gopkg.in/yaml.v3"
 
-	"github.com/livekit/livekit-egress/pkg/errors"
 	"github.com/livekit/protocol/livekit"
 	"github.com/livekit/protocol/logger"
 	"github.com/livekit/protocol/utils"
@@ -72,7 +72,7 @@ func NewConfig(confString string) (*Config, error) {
 	}
 	if confString != "" {
 		if err := yaml.Unmarshal([]byte(confString), conf); err != nil {
-			return nil, errors.ErrCouldNotParseConfig(err)
+			return nil, fmt.Errorf("could not parse config: %v", err)
 		}
 	}
