+++ pkg/pipeline/params/params.go
@@ -119,14 +119,11 @@ func GetPipelineParams(conf *config.Config, request *livekit.StartEgressRequest)
 		default:
 			return nil, errors.ErrInvalidInput("output")
 		}
-
 	case *livekit.StartEgressRequest_TrackComposite:
 		params.Info.Request = &livekit.EgressInfo_TrackComposite{TrackComposite: req.TrackComposite}
 
-		params.AudioTrackID = req.TrackComposite.AudioTrackId
-		params.AudioEnabled = params.AudioTrackID != ""
-		params.VideoTrackID = req.TrackComposite.VideoTrackId
-		params.VideoEnabled = params.VideoTrackID != ""
+		params.AudioEnabled = req.TrackComposite.AudioTrackId != ""
+		params.VideoEnabled = req.TrackComposite.VideoTrackId != ""
 		params.RoomName = req.TrackComposite.RoomName
 
 		switch o := req.TrackComposite.Output.(type) {
@@ -140,10 +137,9 @@ func GetPipelineParams(conf *config.Config, request *livekit.StartEgressRequest)
 			if err := params.updateStreamInfo(o.Stream.Protocol, o.Stream.Urls); err != nil {
 				return nil, err
 			}
-		default:
-			return nil, errors.ErrInvalidInput("output")
 		}
 
+		return nil, errors.ErrNotSupported("track composite requests")
 	case *livekit.StartEgressRequest_Track:
 		params.Info.Request = &livekit.EgressInfo_Track{Track: req.Track}
