+++ pkg/pipeline/pipeline.go
@@ -5,11 +5,13 @@ import (
 
 	"github.com/livekit/livekit-egress/pkg/config"
 	"github.com/livekit/livekit-egress/pkg/errors"
+	"github.com/livekit/livekit-egress/pkg/pipeline/composite"
 	"github.com/livekit/livekit-egress/pkg/pipeline/params"
+	"github.com/livekit/livekit-egress/pkg/pipeline/track"
 )
 
 type Pipeline interface {
-	GetInfo() *livekit.EgressInfo
+	Info() *livekit.EgressInfo
 	Run() *livekit.EgressInfo
 	UpdateStream(req *livekit.UpdateStreamRequest) error
 	Stop()
@@ -28,11 +30,11 @@ func FromRequest(conf *config.Config, request *livekit.StartEgressRequest) (Pipe
 func FromParams(conf *config.Config, p *params.Params) (Pipeline, error) {
 	switch p.Info.Request.(type) {
 	case *livekit.EgressInfo_RoomComposite:
-		return NewCompositePipeline(conf, p)
+		return composite.NewPipeline(conf, p)
 	case *livekit.EgressInfo_TrackComposite:
-		return NewCompositePipeline(conf, p)
+		return composite.NewPipeline(conf, p)
 	case *livekit.EgressInfo_Track:
-		return NewTrackPipeline(p)
+		return track.NewPipeline(p)
 	default:
 		return nil, errors.ErrInvalidInput("request")
 	}
