--- pkg/pipeline/source/clock.go
@@ -1,37 +1,0 @@
-package source
-
-import "go.uber.org/atomic"
-
-// a single clockSync is shared between audio and video writers
-// used for creating PTS
-type clockSync struct {
-	startTime atomic.Int64
-	endTime   atomic.Int64
-	delay     atomic.Int64
-}
-
-func (c *clockSync) GetOrSetStartTime(t int64) int64 {
-	swapped := c.startTime.CAS(0, t)
-	if swapped {
-		return t
-	}
-	startTime := c.startTime.Load()
-	c.delay.CAS(0, t-startTime)
-	return startTime
-}
-
-func (c *clockSync) GetStartTime() int64 {
-	return c.startTime.Load()
-}
-
-func (c *clockSync) SetEndTime(t int64) {
-	c.endTime.Store(t)
-}
-
-func (c *clockSync) GetEndTime() int64 {
-	return c.endTime.Load()
-}
-
-func (c *clockSync) GetDelay() int64 {
-	return c.delay.Load()
-}
