--- test/sdk_test.go
@@ -1,101 +1,0 @@
-//go:build integration
-// +build integration
-
-package test
-
-import (
-	"testing"
-	"time"
-
-	"github.com/stretchr/testify/require"
-
-	"github.com/livekit/protocol/livekit"
-	"github.com/livekit/protocol/utils"
-	lksdk "github.com/livekit/server-sdk-go"
-
-	"github.com/livekit/livekit-egress/pkg/config"
-)
-
-func testTrackComposite(t *testing.T, conf *config.Config, room *lksdk.Room) {
-	testTrackCompositeFile(t, conf, room, "opus", "vp8", []*testCase{
-		{
-			name:       "track-vp8-mp4",
-			fileType:   livekit.EncodedFileType_MP4,
-			filePrefix: "track-vp8",
-		},
-		{
-			name:      "track-opus-only-ogg",
-			audioOnly: true,
-			fileType:  livekit.EncodedFileType_OGG,
-		},
-	})
-
-	testTrackCompositeFile(t, conf, room, "opus", "h264", []*testCase{
-		{
-			name:       "track-h264-mp4",
-			fileType:   livekit.EncodedFileType_MP4,
-			filePrefix: "track-h264",
-		},
-		{
-			name:      "track-h264-only-mp4",
-			videoOnly: true,
-			fileType:  livekit.EncodedFileType_MP4,
-		},
-	})
-}
-
-func testTrackCompositeFile(t *testing.T, conf *config.Config, room *lksdk.Room, audioCodec, videoCodec string, cases []*testCase) {
-	p := publishSamplesToRoom(t, room, audioCodec, videoCodec)
-
-	for _, test := range cases {
-		if !t.Run(test.name, func(t *testing.T) {
-			runTrackCompositeFileTest(t, conf, p, test)
-		}) {
-			t.FailNow()
-		}
-	}
-
-	require.NoError(t, room.LocalParticipant.UnpublishTrack(p.audioTrackID))
-	require.NoError(t, room.LocalParticipant.UnpublishTrack(p.videoTrackID))
-}
-
-func runTrackCompositeFileTest(t *testing.T, conf *config.Config, params *sdkParams, test *testCase) {
-	filepath, filename := getFileInfo(conf, test, "track")
-
-	var audioTrackID, videoTrackID string
-	if !test.videoOnly {
-		audioTrackID = params.audioTrackID
-	}
-	if !test.audioOnly {
-		videoTrackID = params.videoTrackID
-	}
-
-	trackRequest := &livekit.TrackCompositeEgressRequest{
-		RoomName:     params.roomName,
-		AudioTrackId: audioTrackID,
-		VideoTrackId: videoTrackID,
-		Output: &livekit.TrackCompositeEgressRequest_File{
-			File: &livekit.EncodedFileOutput{
-				FileType: test.fileType,
-				Filepath: filepath,
-			},
-		},
-	}
-
-	if test.options != nil {
-		trackRequest.Options = &livekit.TrackCompositeEgressRequest_Advanced{
-			Advanced: test.options,
-		}
-	}
-
-	req := &livekit.StartEgressRequest{
-		EgressId:  utils.NewGuid(utils.EgressPrefix),
-		RequestId: utils.NewGuid(utils.RPCPrefix),
-		SentAt:    time.Now().UnixNano(),
-		Request: &livekit.StartEgressRequest_TrackComposite{
-			TrackComposite: trackRequest,
-		},
-	}
-
-	runFileTest(t, conf, test, req, filename)
-}
